requirements.txt
torch 
torchvision
pillow
omegaconf
hydra-core
